// src/types/schemas/match-action.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type MatchAction = {
  team?: "LOCAL" | "VISITOR";
  action?:
    | "GOAL"
    | "TIMEOUT"
    | "YELLOW_CARD"
    | "RED_CARD"
    | "FAULT"
    | "PENALTY_FAULT"
    | "PENALTY_GOAL"
    | "PENALTY_MISSED"
    | "PENALTY_SAVED"
    | "SAVE"
    | "BALL_LOST"
    | "BALL_POSSESSION"
    | "FASTBREAK_GOAL"
    | "FASTBREAK_SAVE"
    | "FASTBREAK_MISSED"
    | "FASTBREAK_ASSIST"
    | "SHOOT_6M"
    | "END_OF_PERIOD"
    | "ASSIST";
  payload?: unknown;
  [k: string]: unknown;
}[];